Bruno tvOS: release date under titles on the Home New Releases shelf
Extends the premiere-date subtitle (added for the Decades browse shelves) to
the Home "New Releases" spine shelf. BrunoTitleDateContentView is extracted
from BrunoShelfRow into its own internal file so both the browse row and
BrunoShelfView can use it (no behavior change to the Decades shelves).
BrunoShelfView passes it as the PosterHStack label only when
shelf.kind == .recentlyAdded; every other Home shelf keeps the default
TitleSubtitleContentView. premiereDate is a top-level field (present even
under RecentlyAddedLibrary's MinimumFields), filling movies' reserved blank
subtitle line so INV-1's pinned row height is unchanged.
